[refactor]: Use global state management with Pinia
Switched to using Pinia for all state management.
All states now ~stores/index.ts and can be accessed and managed by every component individually and be reactively updated for every other component.

This is a major refactor and will be need to be further bug tested but it looks stable so far and should be ready for prod